Extract Computes, ComputeTasks to a separate module  and add hangar computes to Computes container (OMS-361)
